refactor vim2
