+++ brainDecTool/math/__init__.py
@@ -1,0 +1,1 @@
+from base import corr2_coef
