+++ brainDecTool/vim2/util.py
@@ -29,17 +29,6 @@ def idx2coord(vec_idx):
     coord_y = (vec_idx % (data_size[1]*data_size[2])) / data_size[2]
     return (coord_x, coord_y, coord_z)
 
-def corr2_coef(A, B):
-    """Row-wise Correlation Coefficient calculation for two 2D arrays."""
-    # Row-wise mean of input arrays & subtract from input arrays themselves
-    A_mA = A - A.mean(1)[:, None]
-    B_mB = B - B.mean(1)[:, None]
-    # Sum of squares across rows
-    ssA = (A_mA**2).sum(1)
-    ssB = (B_mB**2).sum(1)
-    # Finally get corr coef
-    return np.dot(A_mA, B_mB.T)/np.sqrt(np.dot(ssA[:, None], ssB[None]))
-
 def node2feature(layer_name, node_idx):
     """Convert node index from CNN activation vector into 3 features including
     index of channel, row and column position of the filter.
